phandinhphu: Responsive giao diện độc giả cho mobile và tablet, thêm 2 biến môi trường WEBSOCKET_SERVER và WEBSOCKET_PORT, fix bug thay đổi ảnh account độc giả

+++ app/controllers/admin/Account.php
@@ -200,7 +200,9 @@ class Account extends Controller
 
             // Xoá ảnh cũ
             if ($account['avatar'] !== $checkAvatar['path'] && $checkAvatar['status'] === 'success') {
-                unlink($account['avatar']);
+                if ($account['avatar'] !== 'uploads/no-image.png') {
+                    unlink($account['avatar']);
+                }
             }
 
             $dataReq['avatar'] = $checkAvatar['path'];
